requirements.txt
Flask==3.0.2
PyJWT==2.8.0
cryptography==42.0.5
pytest==8.0.0
pytest-cov==5.0.0